requirements.txt
python-dotenv~=1.1.1
redis~=6.2.0
aiohttp~=3.12.13